When updating a whole item (which has an item_no, not an id), return to results page with mode=update and not mode=updateItem.
This enables the "update" buttons on the results page for this item number. Apparently, there's no special PHP preprocessing of the HTML for mode=updateItem and that's why no "update" buttons appeared. However, mode=update is the URL parameter when simply entering the update mode, which is what Efrat wanted.