Print the DOCTYPE statement from the PHP code because it has to be at the top of the HTML page but nothing can be printed in the file before session_start() or there is a warning message displayed on the HTML page.
Also, changed the minChars to 2 because it was a bit too heavy to use 1 as the value and also because of a small bug preventing the display of autoComplete suggestions in Hebrew if the user typed a non-letters which are found in the Hebrew name field of the composer (hebrew_composer). The issue is tracked in the Issues tab in Google Code.