In the autocomplete list, put results starting with query (prefix) before results that just contain the query (infix).
This is a bit buggy and it is a bit unclear why not all results are displayed on the list. This is probably a bit unrelated but there's a problem with hebrew requiring "like binary". We need to detect when to use "like" and when to use "like binary".

To fix, consider one of the following:
1) Querying only for prefix: $x like '$q%'
2) Querying for: $x like '$q%' and then querying again for $x like '$%q%' and not $x like '$q%', to get the infix results too after the prefix results