"Found problems? Fix them..." message for copied items was displayed only for the first copied item. After that, if we continued copying that item the message was not displayed. This was a bug.
The problem was, the next time a condition changed $action to "newitem" and there was no entry to the succeeding that was supposed to set up the appearance of the above message.

Now, $action="newitem" also allows the display of that message if an $item_no was given in the URL (because we need to have an item with details that may require fixing).